Decryption now works.
threshcrypt can now decrypt it's encrypted files.

Still not suitable for actual use since it spews key material all over the
screen. :-P

Also did some cleanup (mostly removing C++ style comments and making casts
explicit) to get warning-free builds in gcc with -pedantic -std=gnu99 enabled.

+++ util.c
@@ -20,7 +20,7 @@ void * safe_malloc(size_t size) {
   void *ptr = malloc(size);
   if (ptr == NULL) {
     perror("malloc");
-    fprintf(stderr, "malloc(%zd) returned NULL\n", size);
+    fprintf(stderr, "malloc(%d) returned NULL\n", (unsigned int)size);
     exit(EXIT_FAILURE);
   }
   memset(ptr, 0, size);
@@ -71,4 +71,4 @@ void fill_rand(unsigned char *buffer,
   fclose(devrandom);
 }
 
-// vim: ts=2 sw=2 et ai si
+/* vim: set ts=2 sw=2 et ai si: */
